refactor common components out
